Changed thickness to trackThickness. Changed type from float to int.

+++ NMRangeSlider/MMColorForTrack.h
@@ -9,5 +9,5 @@
 #import <UIKit/UIKit.h>
 
 @interface MMColorForTrack : NSObject
-+(UIImage*)getTrackImageWithColorR:(NSInteger)r G:(NSInteger)g B:(NSInteger)b A:(NSInteger)a Thickness:(float)thickness;
++(UIImage*)getTrackImageWithColorR:(NSInteger)r G:(NSInteger)g B:(NSInteger)b A:(NSInteger)a Thickness:(int)thickness;
 @end
